% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_probabilities.R
\encoding{UTF-8}
\name{conditional_probabilities}
\alias{conditional_probabilities}
\title{Computes the conditional probabilities of a categorical time series}
\usage{
conditional_probabilities(series, lag = 1, categories)
}
\arguments{
\item{series}{A CTS.}

\item{lag}{The considered lag (default is 1).}

\item{categories}{A vector of type factor containing the corresponding
categories.}
}
\value{
A matrix with the conditional probabilities.
}
\description{
\code{conditional_probabilities} returns a matrix with the conditional
probabilities of a categorical time series
}
\details{
Given a CTS of length \eqn{T} with range \eqn{\mathcal{V}=\{1, 2, \ldots, r\}},
\eqn{\overline{X}_t=\{\overline{X}_1,\ldots, \overline{X}_T\}}, the function computes the
matrix \eqn{\widehat{\boldsymbol P}^c(l) = \big(\widehat{p}^c_{ij}(l)\big)_{1 \le i, j \le r}},
with \eqn{\widehat{p}^c_{ij}(l)=\frac{TN_{ij}(l)}{(T-l)N_i}}, where
\eqn{N_i} is the number of elements equal to \eqn{i} in the realization \eqn{\overline{X}_t} and \eqn{N_{ij}(l)} is the number
of pairs \eqn{(\overline{X}_t, \overline{X}_{t-l})=(i,j)} in the realization \eqn{\overline{X}_t}.
}
\examples{
matrix_cp <- conditional_probabilities(series = GeneticSequences$data[[1]],
categories = factor(c('a', 'c', 'g', 't'))) # Computing the matrix of
# joint probabilities for the first series in dataset GeneticSequences
}
\references{
{

  \insertRef{weiss2008measuring}{ctsfeatures}

}
}
\author{
Ángel López-Oriona, José A. Vilar
}
