% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts_plot.R
\encoding{UTF-8}
\name{cts_plot}
\alias{cts_plot}
\title{Constructs a categorical time series plot}
\usage{
cts_plot(series, categories, title = "Time series plot")
}
\arguments{
\item{series}{A CTS.}

\item{categories}{A vector of type factor containing the corresponding
categories.}

\item{title}{The title of the graph.}
}
\value{
The categorical time series plot.
}
\description{
\code{cts_plot} constructs a categorical time series plot
}
\details{
Constructs a categorial time series plot for a given CTS.
}
\examples{
time_series_plot <- cts_plot(series = GeneticSequences$data[[1]][1 : 50],
categories = factor(c('a', 'c', 'g', 't'))) # Constructs a categorical
# time series plot for the first 50 observations of the first  time series in
# dataset GeneticSequences
}
\references{
{

  \insertRef{weiss2018introduction}{ctsfeatures}

}
}
\author{
Ángel López-Oriona, José A. Vilar
}
