% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cts.R
\encoding{UTF-8}
\name{plot_cts}
\alias{plot_cts}
\title{Constructs a categorical time series plot}
\usage{
plot_cts(series, categories, title = "Time series plot")
}
\arguments{
\item{series}{A CTS.}

\item{categories}{A vector of type factor containing the corresponding
categories.}

\item{title}{The title of the graph.}
}
\value{
The categorical time series plot.
}
\description{
\code{plot_cts} constructs a categorical time series plot
}
\details{
Constructs a categorial time series plot for a given CTS.
}
\examples{
time_series_plot <- plot_cts(series = GeneticSequences$data[[1]][1 : 50],
categories = factor(c('a', 'c', 'g', 't'))) # Constructs a categorical
# time series plot for the first 50 observations of the first  time series in
# dataset GeneticSequences
}
\references{
{

  \insertRef{weiss2018introduction}{ctsfeatures}

}
}
\author{
Ángel López-Oriona, José A. Vilar
}
