% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg.R
\encoding{UTF-8}
\name{plot_reg}
\alias{plot_reg}
\title{Constructs the rate evolution graph for a categorical time series}
\usage{
plot_reg(
  series,
  title = "Rate evolution graph",
  linear_fit = FALSE,
  cat_res = NULL,
  ...
)
}
\arguments{
\item{series}{An object of type \code{tsibble} (see R package \code{tsibble}), whose column named Values
contains the values of the corresponding CTS. This column must be of class \code{factor} and its levels
must be determined by the range of the CTS.}

\item{title}{The title of the graph.}

\item{linear_fit}{Logical. I \code{TRUE}, the corresponding least squares
lines are incorporated to the graph}

\item{cat_res}{If this parameter is set to any of the categories of the
series, then the function returns a graph of residuals for the linear model
associated with the corresponding category}

\item{...}{Additional parameters for the function.}
}
\value{
The rate evolution graph.
}
\description{
\code{plot_reg} constructs the rate evolution graph
proposed by \insertCite{ribler1997visualizing;textual}{ctsfeatures}.
}
\details{
Given a CTS of length \eqn{T} with range \eqn{\mathcal{V}=\{1, 2, \ldots, r\}},
\eqn{\overline{X}_t=\{\overline{X}_1,\ldots, \overline{X}_T\}}, and the
corresponding binarized time series, \eqn{\overline{\boldsymbol Y}_t=\{\overline{\boldsymbol Y}_1, \ldots, \overline{\boldsymbol Y}_T\}},
the function constructs the rate evolution graph. Specifically, consider the
series of cumulated sums given by \eqn{\overline{\boldsymbol C}_t=\{\overline{\boldsymbol C}_1, \ldots, \overline{\boldsymbol C}_T\}}, with
\eqn{\overline{\boldsymbol C}_k=\sum_{s=1}^{k}\overline{\boldsymbol Y}_s},
\eqn{k=1,\ldots,T}. The rate evolution graph displays a standard time series
plot for each one of the components of \eqn{\overline{\boldsymbol C}_t}
simultaneously in one graph.
}
\examples{
sequence_1 <- GeneticSequences[which(GeneticSequences$Series==1),]
reg <- plot_reg(sequence_1) # Constructing the rate
# evolution graph for the first time series in dataset GeneticSequences
}
\references{
{

  \insertRef{ribler1997visualizing}{ctsfeatures}

}
}
\author{
Ángel López-Oriona, José A. Vilar
}
