% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_se.R
\encoding{UTF-8}
\name{plot_se}
\alias{plot_se}
\title{Represents the spectral envelope of a categorical time series}
\usage{
plot_se(series, categories)
}
\arguments{
\item{series}{A CTS.}

\item{categories}{A vector of type factor containing the corresponding
categories.}
}
\value{
Returns returns a plot of the spectral envelope.
}
\description{
\code{plot_se} represents the spectral envelope of
a categorical time series
}
\details{
The function represents the spectral envelope of a categorical time series
}
\examples{
plot_se(GeneticSequences$data[[1]],
categories = factor(c('a', 'c', 'g', 't')))
# Representing the spectral envelope for the first series in dataset
# GeneticSequences
}
\references{
{

  \insertRef{stoffer1993spectral}{ctsfeatures}

}
}
\author{
Ángel López-Oriona, José A. Vilar
}
