% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SleepStages.R
\docType{data}
\name{SleepStages}
\alias{SleepStages}
\title{SleepStages}
\format{
A \code{tsibble} with four columns, which are:
\describe{
\item{\code{Value}}{The categorical values of the time series in the dataset.}
\item{\code{Series}}{Integer values indicating the considered time series (there are 62 time series in the dataset).}
\item{\code{Time}}{Integer values indicating the temporal indexes of the observations.}
\item{\code{Class}}{Integer values indicating the class of each time series.}
}
}
\usage{
data(SleepStages)
}
\description{
Categorical time series (CTS) of sleep stages from different
subjects
}
\details{
The column \code{Value} is the concatenation of 62 time series
taking six categorical values (sleep stages). The column \code{Class} is formed
by the integers 1 and 2 indicating that there are 2 different classes in the database. Each class is associated with a sleep
disorder (class 1 refers to nocturnal frontal lobe epilepsy, while class refers 2 to REM behavior disorder).
For more information, see \insertCite{lopez2023hard;textual}{ctsfeatures}.
}
\references{
{

  \insertRef{lopez2023hard}{ctsfeatures}

}
}
\keyword{datasets}
