% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntheticData1.R
\docType{data}
\name{SyntheticData1}
\alias{SyntheticData1}
\title{SyntheticData1}
\format{
A \code{list} with two elements, which are:
\describe{
\item{\code{data}}{A list with 80 CTS.}
\item{\code{classes}}{A numeric vector indicating the corresponding classes associated with the elements in \code{data}.}
}
}
\usage{
data(SyntheticData1)
}
\description{
Synthetic dataset containing 80 CTS generated from four
different generating processes.
}
\details{
Each element in \code{data} is a CTS of length 600 containing three
different categories. Series 1-20, 21-40, 41-60 and 61-80 were generated from
Markov Chains with different matrices of transition probabilities (see Scenario 1 in \insertCite{lopez2023hard;textual}{ctsfeatures}).
Therefore, there are 4 different classes in the dataset.
}
\references{
{

  \insertRef{lopez2023hard}{ctsfeatures}

}
}
\keyword{datasets}
