% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_se.R
\encoding{UTF-8}
\name{plot_se}
\alias{plot_se}
\title{Represents the spectral envelope of a categorical time series}
\usage{
plot_se(series)
}
\arguments{
\item{series}{An object of type \code{tsibble} (see R package \code{tsibble}), whose column named Values
contains the values of the corresponding CTS. This column must be of class \code{factor} and its levels
must be determined by the range of the CTS.}
}
\value{
Returns returns a plot of the spectral envelope.
}
\description{
\code{plot_se} represents the spectral envelope of
a categorical time series
}
\details{
The function represents the spectral envelope of a categorical time series
}
\examples{
sequence_1 <- GeneticSequences[which(GeneticSequences$Series==1),]
plot_se(sequence_1)
# Representing the spectral envelope for the first series in dataset
# GeneticSequences
}
\references{
{

  \insertRef{stoffer1993spectral}{ctsfeatures}

}
}
\author{
Ángel López-Oriona, José A. Vilar
}
