% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinSequences.R
\docType{data}
\name{ProteinSequences}
\alias{ProteinSequences}
\title{ProteinSequences}
\format{
A \code{tsibble} with four columns, which are:
\describe{
\item{\code{Value}}{The categorical values of the time series in the dataset.}
\item{\code{Series}}{Integer values indicating the considered time series (there are 40 time series in the dataset).}
\item{\code{Time}}{Integer values indicating the temporal indexes of the observations.}
\item{\code{Class}}{Integer values indicating the class of each time series.}
}
}
\usage{
data(ProteinSequences)
}
\description{
Categorical time series (CTS) of protein sequences from different
species
}
\details{
The column \code{Value} is the concatenation of 40 time series
taking four categorical values (amino-acids). The column \code{Class} is formed
by integers from 1 to 4, indicating that there are 4 different classes in the database. Each class is associated with a different
family of viruses. For more information, see \insertCite{lopez2023hard;textual}{ctsfeatures}.
}
\references{
{

  \insertRef{lopez2023hard}{ctsfeatures}

}
}
\keyword{datasets}
